Create DeploymentCode
This code takes the possession of the hand and each finger. Then the code reads what hand gesture is being made in front of the camera using OpenCV. An output is sent commands to an Arduino board using Pyfirmata2. The outputs are LED lights. Each light represents that a specific hand gesture is being made.