Rename DeploymentCode to DeploymentCode.py
This code deploys the AI module from the "HandTrackingModule.py". This code uses OpenCV to take in the hand gesture. Then the AI determines the possession of the hand. The code then decides what hand gesture is being made and comes with an output. Once a decision is made, the code sends a command to an Arduino Uno board to turn on a specific LED light that corresponds to that hand gesture. The communication between python and Arduino IDE is made with PyFirmata library.